More opcodes: SLL, ADDIU, SW

+++ qPlayStation/src/qPlayStation.cpp
@@ -95,7 +95,7 @@ void initOpenGL()
     vs = glCreateShader(GL_VERTEX_SHADER);
     fs = glCreateShader(GL_FRAGMENT_SHADER);
 
-    int length = strlen(vertex_shader);
+    int length = (int)strlen(vertex_shader);
     glShaderSource(vs, 1, (const GLchar**)&vertex_shader, &length);
     glCompileShader(vs);
 
@@ -107,7 +107,7 @@ void initOpenGL()
         exit(1);
     }
 
-    length = strlen(fragment_shader);
+    length = (int)strlen(fragment_shader);
     glShaderSource(fs, 1, (const GLchar**)&fragment_shader, &length);
     glCompileShader(fs);
